+++ apps/sanity-cms/schemas/documents/handbook.ts
@@ -1,8 +1,8 @@
 import { BiBookBookmark } from 'react-icons/bi';
 
 export default {
-  name: 'almanac',
-  title: 'Almanac',
+  name: 'handbook',
+  title: 'Handbook',
   type: 'document',
   icon: BiBookBookmark,
   groups: [
@@ -20,32 +20,32 @@ export default {
       name: 'title',
       type: 'string',
       title: 'Title',
-      description: 'Enter the title for the almanac item.',
+      description: 'Enter the title for the handbook item.',
       group: 'metadata',
       validation: (Rule: any) => [
         Rule.required()
           .min(5)
-          .error('The almanac item title should be longer than 5 characters.'),
+          .error('The handbook item title should be longer than 5 characters.'),
         Rule.required()
           .max(96)
           .warning(
             `That's a pretty long title, you might want to provide a shorter title.`
           ),
-        Rule.required().error('You must enter a title for the almanac item.')
+        Rule.required().error('You must enter a title for the handbook item.')
       ]
     },
     {
       name: 'slug',
       type: 'slug',
       title: 'Slug',
-      description: 'The slug determines the canonical organization of this almanac item within the application.',
+      description: 'The slug determines the canonical organization of this handbook item within the application.',
       group: 'metadata',
       options: {
         source: 'title',
       },
       validation: (Rule: any) => [
         Rule.required().error(
-          `The almanac item needs a slug. Create one based on the title with "generate"`
+          `The handbook item needs a slug. Create one based on the title with "generate"`
         ),
       ],
     },
@@ -66,20 +66,20 @@ export default {
       name: 'excerpt',
       title: 'Excerpt',
       type: 'text',
-      description: 'Excerpt for the almanac item.',
+      description: 'Excerpt for the handbook item.',
       group: 'content',
       validation: (Rule: any) => [
-        Rule.required().error('You must enter an excerpt for the almanac item.')
+        Rule.required().error('You must enter an excerpt for the handbook item.')
       ]
     },
     {
       name: 'content',
       title: 'Content',
       type: 'richText',
-      description: 'Content for the almanac item.',
+      description: 'Content for the handbook item.',
       group: 'content',
       validation: (Rule: any) => [
-        Rule.required().error('You must enter content for the almanac item.')
+        Rule.required().error('You must enter content for the handbook item.')
       ]
     }
   ],