+++ apps/sanity-cms/schemas/index.ts
@@ -19,7 +19,7 @@ import genre from './documents/genre'
 import newsletter from './documents/newsletter'
 import snippet from './documents/snippet'
 import resource from './documents/resource'
-import almanac from './documents/almanac'
+import handbook from './documents/handbook'
 
 // Import Objects
 import mainImage from './objects/mainImage'
@@ -35,7 +35,7 @@ import metadata from './objects/metadata'
 import moduleContent from './objects/moduleContent'
 import blockquote from './objects/blockquote'
 import fileDownload from './objects/fileDownload'
-import almanacReference from './objects/almanac-reference'
+import handbookReference from './objects/handbook-reference'
 
 // Navigation
 import navigationSection from './objects/navigation/section'
@@ -80,7 +80,7 @@ export const schemaTypes = [
 	resource,
 	newsletter,
 	snippet,
-	almanac,
+	handbook,
 	// Objects
 	mainImage,
 	richText,
@@ -113,5 +113,5 @@ export const schemaTypes = [
 	barValue,
 	musicSymbol,
 	fileDownload,
-	almanacReference,
+	handbookReference,
 ]