+++ apps/www/src/app/components/navigation/main/nav-content.tsx
@@ -34,8 +34,8 @@ const NavContent = ({ links }) => {
       href: '/contributors'
     },
     {
-      text: 'Almanac',
-      href: '/almanac'
+      text: 'Handbook',
+      href: '/handbook'
     },
     {
       text: 'Contact',