+++ apps/www/CHANGELOG.md
@@ -1,5 +1,18 @@
 # www
 
+## 3.2.0
+
+### Minor Changes
+
+- c5a0cdf: Updated Almanac to Handbook
+
+## 3.1.0
+
+### Minor Changes
+
+- Updated Almanac to Handbook
+- 7ddaad9: Added a newsletter signup page
+
 ## 3.0.0
 
 ### Major Changes